+++ src/main/java/com/ingsw/flyingdutchman/model/mo/Product.java
@@ -6,9 +6,9 @@ import java.sql.Blob;
 public class Product {
     private Long productID;
     private String description;
-    private Integer min_price;
-    private Integer starting_price;
-    private Integer current_price;
+    private Float min_price;
+    private Float starting_price;
+    private Float current_price;
     private Blob image;
     private Boolean deleted;
     private Category category;
@@ -18,12 +18,12 @@ public class Product {
     public void setProductID(Long productID){this.productID = productID;}
     public String getDescription(){return description;}
     public void setDescription(String description) {this.description = description;}
-    public Integer getMin_price(){return min_price;}
-    public void setMin_price(int min_price){this.min_price = min_price;}
-    public Integer getStarting_price(){return starting_price;}
-    public void setStarting_price(int starting_price){this.starting_price = starting_price;}
-    public Integer getCurrent_price(){return current_price;}
-    public void setCurrent_price(Integer current_price){this.current_price = current_price;}
+    public Float getMin_price(){return min_price;}
+    public void setMin_price(float min_price){this.min_price = min_price;}
+    public Float getStarting_price(){return starting_price;}
+    public void setStarting_price(float starting_price){this.starting_price = starting_price;}
+    public Float getCurrent_price(){return current_price;}
+    public void setCurrent_price(float current_price){this.current_price = current_price;}
     public Blob getImage(){return image;}
     public void setImage(Blob image){this.image = image;}
     public Category getCategory() {return category;}
