+++ src/main/java/com/ingsw/flyingdutchman/model/mo/Threshold.java
@@ -4,7 +4,7 @@ import java.sql.Timestamp;
 
 public class Threshold {
     private Long ThresholdID;
-    private Integer price;
+    private Float price;
     private Timestamp reservation_date;
     private User user;
     private Auction auction;
@@ -15,8 +15,8 @@ public class Threshold {
     public void setAuction(Auction auction){this.auction = auction;}
     public User getUser(){return user;}
     public void setUser(User user){this.user = user;}
-    public Integer getPrice(){return price;}
-    public void setPrice(int price){this.price = price;}
+    public Float getPrice(){return price;}
+    public void setPrice(float price){this.price = price;}
     public Timestamp getReservation_date(){return reservation_date;}
     public void setReservation_date(Timestamp reservationDate){this.reservation_date = reservationDate;}
 }
